Allow for absence of controller attribute in owner ref
Pods with ownerReferences that do not contain 'controller' raise a
KeyError when trying to find the root owner in order to recycle the
deployment.

It seems that the 'controller' attribute does not appear in
ownerReferences of pods that are not running, e.g. *-deploy pods that
were not immediately deleted upon termination of the pod.  Therefore it
seems safe to assume pods without the 'controller' attribute in
ownerReferences can be skipped for recyling.