Fix "Request Entity Too Large"
The API was expecting a List of all imagemanifestvuln from a given
cluster, but that list is just way too big.

Now the API expects each imagemanifestvuln object individually. This
patch is the client implementation for that. Now we loop all the items
in the big imagemanifestvuln list and post them one by one.

Signed-off-by: Amador Pahim <apahim@redhat.com>