Main loop: drop subprocess
This patch will drop the subprocess call from the main loop (used to run
the integrations on kubernetes) in favor of a direct call to the Click
entrypoint.

With that, both the Main lopp and the integration execution will happen
in the same Python session, meaning we can more data from the
integration executiuon in the Main loop.

This sets the foundation for instrumenting the integrations to  report
the reconciliation time or any other executions attributes we want to
track.

Signed-off-by: Amador Pahim <apahim@redhat.com>