Replace vault_client module by VaultClient class
The vault_client module has is initialized globally, making the access
token to expire after some time.

Now that we run integrations as part of the main loop session, we have
to stop doing the global init and make sure we create a new instance of
the client for each execution.

The vault_client module will be deleted later, when the secret_reader
stops using it.

Signed-off-by: Amador Pahim <apahim@redhat.com>