utils.threaded: call pool.join()
This gives us a sync point before we continue and do something else. The
down side is that a stuck thread (hypothetical scenario) will make the
integration stuck. But the situation before that was that a stuck thread
would hang forever. Out of the two, making the integration stuck is the
safer.

Signed-off-by: Amador Pahim <apahim@redhat.com>