Add prometheus-rules-validator integration (#1265)
Part of https://issues.redhat.com/browse/APPSRE-1146

It is debatable that this is actually an integration, as it will only
run on pr-check. But the code made sense here, as all the utils are here
and I would have needed to copy too much code elsewhere.

The complication here has been to be able to query qontract-server for
PrometheusRules. This has been done to just resolve the templates we
needed, reducing the run time (in my laptop) from 1m20 to 10s. Since
this is an integration that will run with every namespace change (see
app-interface related PR) and we're committed to reduce the PR check
time, this is important.

Related PRs:

* App-interface schema
  * https://gitlab.cee.redhat.com/service/app-interface/-/merge_requests/13366

* Add schema to resource in app-interface bundle
  * https://github.com/app-sre/qontract-validator/pull/25
  * https://github.com/app-sre/qontract-validator/pull/29

* Support schema in resources in qontract-server
  * https://github.com/app-sre/qontract-server/pull/100

* Avoid checking schema in resource graphql queries:
  * https://github.com/app-sre/qontract-reconcile/pull/1260

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>