Added prometheus tests to prometheus validator (#1315)
Rule tests have been added in the same integration as the validations as
we needed the actual rule contents to run the tests, hence separating it
would have meant to fetch the rules twice. The integration has been
renamed to tester since test is more generic as it  includes validation.

Since rules can be templates, this means rules can have a different
shape for different namespaces. This force us to run the tests in every
namespace the rules appear. To make things simpler, I have limited the
number of rules per test to one as allowing for multiple rules
complicated significantly the run of the tests. If we see this is a
limitation, we can revisit this decision.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>