New integration to check alertmanager config
The approach is not very satisfactory, as we need to know the file where
the config is located. The other approach would be to define a schema
for the alertmanager config in app-interface, but it looks like
duplicating the job as we're going to check the schema anyway. As we
only have one alertmanager config this looks enough for the moment but
we may want to define a very lax schema in app-interface in the future
to better identify these files.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>