Adds thread that refreshes the vault token
The vault token has a TTL. Once that TTL is has expired, when trying to
access vault secrets it will return a permission error, as the token is
no longer valid.

Our code is currently reactive, we watch when that happens, we refresh
the vault token, and try again.

That approach is not optimal as it requires handling the same error in
multiple places. We want to proactively refresh the token before it
expires in order to avoid the problem altogether.

Note that this is required because our integrations are long running and
our goal is to exceed the TTL.

Related MRs:
https://github.com/app-sre/qontract-reconcile/pull/1414
https://github.com/app-sre/qontract-reconcile/pull/1480