Do not handle SecretAccessForbidden
We were previously assuming SecretAccessForbidden was a symptom of token
expiration. Therefore we were catching that exception, refreshing the
token, and raising a SecretNotFound to get the code to retry the
operation.

With the proactive token refresh we don't need to handle this. If the
vault client raises SecretAccessForbidden it's now likely a problem with
the RBAC of the client, so we don't want to hide it.