Use the CIS .metadata.name as the target tag
The tag in the releaseImage has been changed from a tag to a digest.
This has caused the mirroring of the ocp-release repo to use the digest
as the tags.

Instead of using the digest as the tag, we are changing the logic to
pick up the CIS name as the tag name for destination ocp release.