[APPSRE-3325]Introduce code coverage in tox configs (#1597)
* Add a minimum coverage file

* Adapt tox configuration to code coverage

Add a section for coverage reports, bump pytest version and add
dependency on the pytest-cov plug-in.

* Disable broken test_make

A test that runs during normal development (say, a `pytest`) run must
not worry about whether there are uncommitted changes. Of course
during normal development there will be uncommitted stuff!

* Ignore coverage database

It's yet another build artifact we don't want to carry in the repo

* Report coverage gaps as well

Add `--cov-report=term-missing` to the execution.

* Allow test_make to run under Jenkins

If the `HUDSON_HOME` environment variable exists, it means we are
running a CI test and indeed we need this test executed to confirm the
integrity of the repository.