[APPSRE-3369]Improve error handling and reporting in requests_sender (#1645)
If, instead of relying on return values we use exceptions, we can
pass a lot of context to the part of the stack best suited to handle
it or to produce the most useful reporting.

So, I have replaced Popen with run in gpg_encrypt, so that I can
get such an exception while removing code. I have not updated
gpg_key_valid since I'd need to update its callers and they are out
of the scope of this task.

I have also replaced some calls to defer with a more Pythonic
context manager. In this case I have adapted gpg_key_valid, for
consistency and because it doesn't change its behaviour or API. This
removes some code, more notably two imports.

Finally, I have tweaked the integration's run to keep the happy path
uninterrupted - that's the beauty of exceptions. We also guarantee
that we perform as much work as possible, reporting failures but
continuing processing. The dry_run flag now lets us get the encrypted credentials, skipping only the state update and email delivery. This way the flag becomes even more useful.