Report failed repo in saasherder.get_sha (#1676)
* Handle SHA-related problems

When we try to fetch a non-existing commit from the Github API, we get
an exception that prints the offending commit but not the
repository.

In the context of integrations, when we are processing many different
repositories, we don't know which of them is actually giving problems.

This also crashes the entire integration.

Instead, we now capture the exception in `get_moving_diff_saas_file`
and skip to the next target in that `resourceTemplate`.  When logging the GithubException, we present not
only its contents but also the name of the SaaS file and the
repository.