Import the aws_api module
The way Python handles imports makes it impossible for testslide to
reach or to mangle properly classes whose constructor has been mocked
out if they are imported into a different namespace. Instead, we
import the module and reach for the class in its own namespace.

It is the same reason for which with `unittest.mock` we have to change
our mock paths depending on imports:

```python

from foo import Bar

b = Bar()

with patch.object(production.module, 'Bar'):
```

or

```python
import foo
b = foo.Bar()

with patch.object(foo, 'Bar'):
```