ldap-users: Use thread safe client strategy, use manual binds (#1738)
This PR updates the `ldap_users` integration to support thread safety. Currently, `ldap_client.py` is invoked in a multi-threaded fashion during reconciliation. However, the previous implementation of the ldap3 package did not include provisions for thread safety when interacting with an LDAP server. See the note in the ldap3 documentation [here](https://ldap3.readthedocs.io/en/latest/index.html).

The current usage of the ldap client is not particularly sophisticated - the integration simply checks with the ldap server to determine if a user record exists, but does nothing with the returned data and only operates on the boolean status variable. The thread safe connection returns a tuple instead of assigning found data to its object properties. Therefore, I ignore the additional return values and keep the integration (`ldap_users.py`) working as written.

Changes:

* ldap client is moved from global scope to function scope
* connection uses `client_strategy=SAFE_SYNC`
* connection is manually bound and is now returned from init, instead of assigning to global _client
* during usage of ldap connection, client bind status is checked before proceeding, if not bound, `client.bind()` is executed
* after each search operation, the client is unbound

This integration ran on stage for 3-4 hours on Friday with no socket close errors reported.