Use exceptions to communicate errors on each cluster
The go-like style of passing an error code in the return tuple is
highly problematic in Python:

* The callers may mistake the order of the tuple, and since
  many standard classes can be interpreted as booleans, the code will
  still work... giving bad results.
* The lack of type hints makes the errors above very hard to catch,
  and in many circumstances, simple unit tests won't find the issues.
* Our underlying libraries still raise exceptions, so we have to
  duplicate error handling efforts.

Instead, we now raise exceptions. Together with a few configurations
known to cause exceptions and crashes in our calls, we can achieve the
desired goal of making consitent progress despite of errors without
adding any specific code.

Refs APPSRE-3441