fix openshift_saas_deploy_trigger_cleaner PipelineRun sort order
This integration cleans up old PipelineRuns but keeps the newest `minimum` amount of runs
and additionally all runs that are younger than `days` days.

There was a bug sorting the PipelineRuns in the wrong order (from oldest to newest) which then resulted in the `minimum`cut-off producing the `minimum` OLDEST PipelineRuns to b e kept.

There was another bug in the `within_retention_days` function which compared date differences and not hours passed (the actual timedelta), which made python say that everything from yesterday was older than 1 day even though it was just, say, 5 hours old.

Co-authored-by: Mayank Shah <mashah@redhat.com>
Co-authored-by: Nico Schieder <nschiede@redhat.com>
Signed-off-by: Josh Gwosdz <jgwosdz@redhat.com>