Refactor terraform_vpc_peerings.build_desired_*state (#1717)
Refactor all the state gathering to a style that is both more pythonic and better suited for concurrency:

First, we split each of the iterations in the loops to their own methods. Then, we replace in these iterations the go-style of returning a value and an error code to just return the value and raise an exception. We don't carry it all the way to the run method just yet, but doing so would remove a fair amount of code and `sys.exit`s.

Finally, we add unit tests for all this refactoring.

This has been tested by running the integration with `--dry-run --print-only` both in master and in this branch. The output is identical.