Fix Slack add_user_to_usergroup actions to handle rate-limiting and errors better
Before this change, the add_user_to_usergroup actions were not catching errors,
in particular those caused by rate-limiting. This change provides a stopgap by
retrying when rate-limited and logging when other errors have occurred.

Some of the other methods in the SlackApi client do not handle retries/failures
properly either. Going down the road of fixing them all will result in significant
refactoring. That time would be best spent investigating the new Slack API client
which has handling for retries and proper exceptions built-in. This method was fixed
because it appears to be one of the more time-sensitive actions.