Improve error handling in quay-permissions (#1816)
* [quay-api]Remove internal exception class

Instead, we rely on the standard HTTPError, which stores the entire
response object and allows the caller to inspect it and do better
reporting. In particular, it can list the failed URL so that we can
understand what resource we are trying to access that is problematic.

* Use standard raise_for_status on requests errors

It gives a better exception, and is less code.

* [quay-permissions]Improve exception handling

We may fail to get a team's permissions too. Also, if we use an
HTTPError we'll get way more details about what is going wrong.

* Adapt unit tests

* Fix unit tests that get in the way

The gitlab housekeeping tests don't behave correctly due to time zone
issues.

* Improve docstrings