Raise KeyError on ResourceInventory.add_desired
We were silently dropping any desired resources on a namespace that
doesn't expect it. This made it terribly difficult for
openshift_resources_base to report to users that it was skipping some
of their resources. And openshift_resources_base was expecting such an
exception to do the reporting.

So, we raise the exception and suppress it in the places that didn't
care about it in the first place.