Fix handling of managedResourceTypes (#1851)
* Add type hints to init_specs_to_fetch

They make following the code a lot easier.

* Raise exception on unexpected failure modes in oc.apply

If there is a new kind `Invalid value:` error when running `oc.apply`
the current code ignores it silently.

We add a fallback exception to discover such situations.

* Raise KeyError on ResourceInventory.add_desired

We were silently dropping any desired resources on a namespace that
doesn't expect it. This made it terribly difficult for
openshift_resources_base to report to users that it was skipping some
of their resources. And openshift_resources_base was expecting such an
exception to do the reporting.

So, we raise the exception and suppress it in the places that didn't
care about it in the first place.

* Log skipped managedResourceNames in openshift_base

I've had to refactor the namespaces part of init_specs_to_fetch so
that we can iterate by resources instead of by resource types.

It results in some more loops, but less nesting and logs become
possible. The added unit tests pass both on current master and here.

* Report skipped resources in saasherder