Fix template inclusion during packaging
This issue was discovered after the move to installing with pip
install instead of setup.py install. It was discovered that the
template files were missing. The templates were never included in
the correct location in the wheel distribution, only when q-r was
installed locally via setup.py install. This fix will ensure that
they are included in the expected directory for both local installs
and wheel distributions.

As per the setuptools docs, package_data is more appropriate because
these template files are closely related to the package implementation
and they are expected to be in the package directory.