Check for missing bucket with head_bucket during State object creation
The s3 head_object method is not yielding consistent results for missing
buckets (403 vs 404). Therefore the check for bucket existance has been
implemented with the head_bucket endpoint. In order to not hit that endpoint
too often, the check has been moved to the constructor of the State class.
If the bucket does not exist, the State object creation fails with a
StateInaccessibleException.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>