Add detailed boto3 ClientError handling in State.exists (#1904)
Add detailed boto3 ClientError handling in State.exists

Distinguish between 404 errorcodes for missing key and others that make 
the bucket inaccessible and the results of exists() unknown.

The s3 head_object method is not yielding consistent results for missing
buckets (403 vs 404). Therefore the check for bucket existence has been
implemented with the head_bucket endpoint. In order to not hit that endpoint
too often, the check has been placed in the constructor of the State class.
If the bucket does not exist, the State object creation fails with a
StateInaccessibleException.

Refs APPSRE-3889

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>