Log skipped managedResources on openshift_resource_base (#1901)
I've had to refactor the namespaces part of init_specs_to_fetch so
that we can iterate by resources instead of by resource types.

It results in some more loops, but less nesting and logs become
possible. The added unit tests pass both on current master and here.

When managedResourceNames, managedResourceTypeOverride or
managedResourceTypes are empty, the graphql query still returns the
field, but its value is None.