Skip clusters with empty serverUrl in openshift-(users|groups) [APPSRE-3866]
When a new cluster is added, it has an empty serverUrl for some period
of time, which causes openshift-(users|groups) to crash.

A longer-term solution is discussed briefly in APPSRE-3866. This fix
would require raising an exception when server is empty in OCNative.
That will be handled separately as this only appears to be an issue
for two integrations at the moment. A comment was added to OCNative
to make the situation clear for the next person who comes across this
in case the refactor isn't prioritized.