Prevent auto promoter from reverting changes when processing multiple promotion per saas file
When the autopromoter processes more than one target-promotion in a saas file
at the same time, the second promotion reverts the first one.

walkthrough:
* while processing the first promotion, the saas file from the master branch
  is loaded, the target ref is updated and a commit to the temporary branch
  is created
* when another promotion for the same saas file is processed, the saas file
  is loaded again from the master branch, which does not contain the change
  from the first promotion. then the change from the promotion is applied and
  a commit is added to the temporary branch. this commit effectively reverts
  the change from previous commits in the temporary branch to the saas file.

the behaviour can be prevented by reading the saas file from the temporary
branch. for the first promotion to process, this does not make a difference,
because the temporary branch has just been created from master. for all
consecutive commits, it is the right thing to do, to prevent the revert and
build upon the previous changes.

Ref APPSRE-4001

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>