Ensure unique auto promoter MR title
The title of auto promoter MRs have been static, which resulted in dropped
promotions because MergeRequestBase prevents the creation of MRs with
duplicate names.

This has been part of the problem for APPSRE-4001 where the logs of
gitlab-mr-sqs-consumer showed multiple promotions close after one another
for different saas files, and one beeing dropped because of the
duplicate title.

Ref APPSRE-4001

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>