adding pragma no cover
Most of what this code does is call the 'dyn' external dependency. This is a very difficult dependency to test as most of the objects and constructors calls the Dyn API and introduce various side effects in the constructor itself. This is a deprecated module which is very unlikely to change before the service is retired in 2023.

Because of this I don't think there is a lot of value in testing it and thus I am ignoring most of it in test coverage