Fix cases where OCLogMsg wasn't being logged (#1955)
* Fix cases where OCLogMsg wasn't being logged

This fix includes a kwarg for listing all clusters in OC_Map, whether
the value is an OCLogMsg or not, in a way that doesn't break the
existing API. This only appears to be broken for a couple of
integrations at the moment which relied on getting a full list of
clusters to be able to reports errors in the mapping. The concept
of OCLogMsg seems error-prone on the surface. It's possible that
the errors could be logged when they occur instead, but this would
be a more involved refactor. This fix may be all that is needed
for now.

* Add tests for OC_Map.clusters()

* Add missing type hint