Fix storage updates performed by the ocm-clusters integration (#1979)
app-interface uses the 'storage' key and OCM uses the 'storage_quota'
key. Using the wrong key to parse the a-i spec was result in discarding
any storage changes. There are some other changes that can be made here
to ensure that an empty cluster spec patch doesn't occur silently, but
that will be handled in a separate PR.