Fail when namespaces list resource names or overrides of unmanaged types (#1986)
* Re-add logging for skipped resources ininit_specs_to_fetch

* Add StateSpec for managed resource types without instances

If we have a namespace that lists `managedResourceTypes: ` but then
some of those are not listed as `managedResourceNames` they may still
be listed as `openshiftResources`. We want to return StateSpec
instances for those.

* Raise exceptions if there are managed resources of unlisted types

If there is a managedResourceName or a managedResourceTypeOverrides
for a kind not listed in managedResourceTypes that is a bad
configuration and we must make the integration fail on it.