Refactor canonicalize_namespaces
Its output is meant to be passed to init_specs_to_fetch, which
provides with a way to override the managed resource types on its
own. The previous mixture of styles made it impossible to decide if a
mismatched managed resource type was a misconfiguration or an
integration being explicit about its needs.

Finally, the style was a bit confusing - copying a list to then remove
linearly some items is slow and a lot of code.

Now we can apply vault secrets (among other resources) while reporting
misconfigurations related to openshift-resources.