Fail again when resource names or overrides have unexpected kinds (#1993)
* Fail when namespaces list resources of unmanaged types

Before fixing its bug.

This reverts commit 7e127091d99723f38c6dc7807ad0c48a9f77eaa8.

* Do not crash integrations that override managedresource types

In those cases, we assume the integration knows what it's doing and
it's safe to skip any listed but not managed resources.

* Move some test data to fixture files

So that we can reuse them on different openshift*base tests

* Refactor canonicalize_namespaces

Its output is meant to be passed to init_specs_to_fetch, which
provides with a way to override the managed resource types on its
own. The previous mixture of styles made it impossible to decide if a
mismatched managed resource type was a misconfiguration or an
integration being explicit about its needs.

Finally, the style was a bit confusing - copying a list to then remove
linearly some items is slow and a lot of code.

Now we can apply vault secrets (among other resources) while reporting
misconfigurations related to openshift-resources.