terraform --print-to-file handle output from multiple accounts
the --print-to-file of our terraform integrations result in the contents managed by terrascript to be dumped to a file. currently, only the contents of the last handled account is present in the file, because the dump loop keeps overwriting the same file over and over again.

this change makes sure that the file will contain the tf information from all accounts, delimited by `##### account name #####`

Ref: https://issues.redhat.com/browse/APPSRE-4421
Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>