[tf-resources] remove RDS/elasticache password lookup fallback in secrets
*motivation for this change*
the `determine_db_password` function in terrascript_client is used as
a fallback mechanism to extract RDS or elasticache credentials from
kubernetes secrets, when the respective outputs can't be found in the
terraform state file. this secret lookup mechanism was useful for tf
resource takeover of previously non-tf managed resources, since in those
cases the tf statefile did not contain any outputs for those credentials.

*why do we need this change*
we don't need this fallback mechanism anymore and it is the reason
an OC_Map is dragged through the terraform_client call stack. while
this isolated change does not seem to have a purpose on its own, it
is a step forward into refactoring terraform_resources, terrascript_client
and terraform_client.

Part of https://issues.redhat.com/browse/APPSRE-4424

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>