Enable untyped def checks in Mypy (#2205)
Mypy only checks typed methods by default. This can be a sane default
for teams starting out with types. Now that we're enforcing types, it
makes more sense to enable a stricter mode so that Mypy will always
check functions for errors. There are >80 files that have errors that
Mypy would detect currently, but they were being ignored because there
weren't type hints. This should stop that behavior for new code and for
files that don't have existing errors.