Cleanup Mypy errors for amtool, slack_api, and secret_reader
The error in slack_api is a pretty cryptic case of __getitem__ overloading.
Rather than making changes to the code to avoid this, which will require
more cycles of manual testing, a single ignore is being added, which is better
than an ignore for the whole module. Once all modules are being checked for
untyped defs, we could start chasing these cryptic cases that may require
code fixes.

Related to the amtool error, you cannot catch the broad Exception and then
access attributes like stdout that only exist on certain types of exceptions.