Add some type hints and docs for Terrascript/Terraform related code
This isn't a huge change, but hopefully it will help reduce the time
to understand the code and catch type-related issues that could be
introduced in the future.

Note that all cases of using concrete types in parameter definitions
is probably intentional. In these cases we're probably passing the
inputs as outputs directly, which makes it impossible to use abstract
types while returning concrete types. Casting to the expected return
type could have been used, but I went out of my way to avoid changing
the code due to limited test coverage.