controller managed labels per kind
cluster controllers that are labeling resources break our drift detection
for openshift_resources. this change introduces a way to declare
which labels on which kind are supposed to be ignored because
their existance or values are managed by a cluster controller.

e.g. ACM (Advanced Cluster Manager) controllers put dynamic labels on
the `ManagedCluter` CRs, e.g. the generated ID of a cluster or the
detected Openshift `version` running on the cluster.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>