Skip Terraform apply when the only RDS update would be engine_version
RDS engine_version changes are unique because a tenant may want to change
this value, but allow RDS to upgrade the instance during the next maintenance
window. This window could be up to 7 days away, which in the meantime would
cause Terraform to apply the change every time that the integration runs (as
well as log the change, which is more annoying than harmful). Unnecessary
apply operations will result in state changes and any instances of
terraform-resources running will need to restart (think PR checks).