Skip Terraform apply when the only RDS update would be engine_version (#2254)
* Skip Terraform apply when the only RDS update would be engine_version

RDS engine_version changes are unique because a tenant may want to change
this value, but allow RDS to upgrade the instance during the next maintenance
window. This window could be up to 7 days away, which in the meantime would
cause Terraform to apply the change every time that the integration runs (as
well as log the change, which is more annoying than harmful). Unnecessary
apply operations will result in state changes and any instances of
terraform-resources running will need to restart (think PR checks).

* Switch AWS API calls to use the AWSApi class and other init refactoring

* Change log line from warning -> debug

* Add type hints that would have caught bug in qontract-cli terraform_users_credentials

* Remove unused argument and make docstring more accurate