use resource specs to write terraform output to vault
this commit starts using resource specs and the embedded secret to write the terraform output to vault. previously the resourceinventory was used for that. upcoming changes will introduce terraform output formats, that will change the way the kuberrnetes secret resource are rendered, and therefore the value in the resourceinventory might not contain the expected and required values for vault anymore.

using the `secret` field of the resource spec ensures that the output from terraform will land in vault as it is required.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>