add TerraformResourceSpec dataclass
this dataclass holds data and context for terraform resources, namely:
* the raw resource specification from qontract-server
* the raw namespace and cluster data from qontract-server
* the output data from the last terraform run

at this point, the wrapped data itself is just raw qontract-server data and not represented as dataclasses themselves. terrascript-client still requires this raw data to act on, so wrapping graphql data for tf resources itself does not add value right now. for fields that are commonly used, TerraformResourceSpec provides properties to access the underlying dicts and provides type hints to them. this also allows to declare common calculated fields in a central place (e.g. the frequently used output prefix).

The TerraformResourceIdentifier is a companion dataclass that can be used as a dictionary key to identify a TerraformResourceSpec by its defining properties: identifier, provider, account
the implementation of this identifier dataclass might change in the future when we start using terraform with other providers than AWS.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>