remodel the StateSpec class to a dataclass with subclasses for current and desired
depending if the StateSpec class was used for current or desired spec, half of the fields were empty. the resource field was even used differently with different type and content, making it harder to reason about the contents of a StateSpec object than necessary.

this commit pulled the common fields for current and desired StateSpec into a BaseStateSpec dataclass and introduced two subclasses for each usecase. this way, these classes and their usage context are easier to test.

exisiting tests have been migrated to pytest and new tests have been introduced to ensure correctness.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>