[terraform-resources] refactor desired state and vault handling (#2331)
Refactor the desired state and vault handling for terraform resource output. This change introduces additional unittesting and typehints, to prepare for future changes like terraform output formats and shared terraform resources.

## TerraformResourceSpec dataclass
This refactor introduces a dataclass `TerraformResourceSpec` that holds data and context for terraform resources, namely:
* the raw resource specification from qontract-server
* the raw namespace and cluster data from qontract-server
* the output data from the last terraform run
this way a resource spec can be passed around with its close context. also frequently used fields are defined as properties with typehints.

## Desired state management
The population of the `ResourceInventory` and creation of the kube `Secret` have been extracted from `terraform_client.py` to `TerraformResourceSpec`. This way the terraform client can concentrate exclusively on terraform matters and is not concerned with close kubernetes matters like `Secret` management. terraform-client keeps the responsibility of parsing terraform output, and making the per-resource grouped outputs available as the `secret` field of the `TerraformResourceSpec` - see `populate_terraform_output_secrets`

This change also makes sense since the required context to build a `Secret` is now available within `TerraformResourceSpec`. Also, implementation of [terraform output formats](https://github.com/app-sre/qontract-schemas/pull/119) will be contained within the `TerraformResourceSpec`.

## Use TerraformResourceSpec as source for Vault population
The secret embedded in `TerraformResourceSpec` is used to write the terraform out data to vault. Previously the `ResourceInventory` was used for that. Upcoming changes will introduce [terraform output formats](https://github.com/app-sre/qontract-schemas/pull/119), that will change the way the kuberrnetes secret resources are rendered, and therefore the value in the `ResourceInventory` will not necessary the format anymore that needs to be written to Vault.

## Typehints & Tests
Typehints and tests were introduced where code was changed or where it was necessary for mypy to be happy.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>