automatic integration sharding based on AWS accounts in app-interface
implementation details: delegate the handling for shards to the `IntegrationShardManager` which knows how to react to different strategies and builds sharding specs for each integration.

a sharding spec is part of the generated `values` file within `integration_spec.shard_specs` which is passed to the helm chart.

a sharding spec consists of the following fields:

* `shard_id`, `shards` - the backwards compatible fields for the `SHARD_ID` and `SHARDS` env vars, specifying which shard within a shard set a deployment will represent.
* `shard_name_suffix` - this i used to influence the `name` of the deployment/statefulset and the `app` label and `matchSelector`
* `extra_args` - contains the aggregation of an integrations global `extraArgs` and the shard specific extra args

this is the default and backward compatible strategy. an integration is running per default with the `static` sharding strategy with 1 shard. the `IntegrationsManager` creates one shard without any `shard_name_suffix`. if `spec.shards > 1` a respective number of `shard_specs` with proper `shard_name_suffixes` (e.g. -0, -1, -2, ...) is created

this strategy creates a shard for each AWS account in app-interface. if an integration is disabled on an account, no shard is created (or an existing one is deleted). the AWS account name is also used as the `shard_name_suffix`, which results in deployments/statefulsets being named `qontract-reconcile-$integration-$account`. this strategy also expects an integration to support an extra arg `--account-name` and provides it to `shard_spec.extra_args`.

the Helm chart expects a list containing at least one `shard_specs` within an `integration_spec`.

part of https://issues.redhat.com/browse/APPSRE-2963
design doc: https://gitlab.cee.redhat.com/service/app-interface/-/blob/master/docs/app-sre/design-docs/qontract-operator-sharding.md

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>