Added rule length check to prometheus rules tester
Prometheus operator establishes a limit on the size of a single
PrometheusRule based on the maximum size of a ConfigMap. From that
moment on it will stop processing new rules.

Since we have rules that are heavily templated via queries, it is not
that uncommon that we get to that limit. This will protect us from that
to happen.

Part of APPSRE-5640

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>