[integration-manager] shard_id as identifier for workload buckets
shard_id has been an integer to identify a bucket of workloads within a collection of shards. since a hashing strategy is used to distribute workload based on the number of buckets, shard_id does not necessarily represent a stable bucket of workload but needs the `shards` label along with it.

this change seeks to make shard_id a true identifier for a workload bucket by allowing it to be one of the following
* shard-nrOfShards, e.g. 2-18
* sharding key, e.g. an aws account

the first option will be enabled by standard sharding (`static` strategy) while the second one will be served by other sharding strategies like the `per-aws-account` strategy.

alerting rule grouping based on `integration` and `shard_id` will be unaffected by this change (maybe short hickups during an initial rollout). also, the extra label feature, used exclusively by terraform-resources to accomodate for the wrapper functionality, will be obsolete.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>